+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # New
 - `harpy view envs`: print versions
 - CLI validations for `--hpc`/`-H` to check if the plugins necessary for the configuration are installed
+- `harpy preprocess meier2021` now uses `dmox` v0.3, introducing the `--use-stitch-base` logic via `--stitch`
 
 # Fixes
 - `harpy view envs`: simpler logic and print diagnostic text if empty